add more into impl, refactor
